+++ memo/ChangeLog.md
@@ -80,6 +80,7 @@
 - main: fix a bug that unset `IFS` is not correctly restored `#D1489` 808f6f7
   - edit: fix error messages on accessing undo records in emacs mode (reported by rux616) `#D1497`  61a57c0 e9be69e
 - canvas: fix a glitch that SGR at the end of command line is applied to new lines `#D1498` 4bdfdbf
+- syntax: fix a bug that `eval() { :; }`, `declare() { :; }` are not treated as function definition `#D1529` 0000000
 
 ## Compatibility